[Sidebar] Changed behavior on tablet and desktop
